pyproject.toml
[project]
name = "download-odds-api-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "matplotlib>=3.7.5",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
]
